initial support for 64-bit builds of Soundflower Bed
git-svn-id: https://soundflower.googlecode.com/svn/trunk@87 0051ed44-974d-0410-95fd-975f363cc5de

+++ SoundflowerBed/AppController.mm
@@ -14,7 +14,7 @@ Boolean startOnAwake = false;
 void	CheckErr(OSStatus err)
 {
 	if (err) {
-		printf("error %-4.4s %ld\n", (char *)&err, err);
+		printf("error %-4.4s %i\n", (char *)&err, (int)err);
 		throw 1;
 	}
 }
